Made the text of "Reversing Axes" (possibly) clearer.
Changed the rule to "i.e. not in the classic 'PC orientation matches cardinal directions'." as it is generally more digestible and does not only apply to Europeans as the previous statement implied.
"reversing the column" is misleading as that could mean ordering it the other way around (first row becomes last and vice versa). 
Reworded that to "flipping the sign of the column's contents." 
See new text.